util: hoist the fuse.conf parsing and setuid mode enforcement code
Move all the code that parses fuse.conf into a separate file in util/ so
that fuservicemount can read the same file, then add the security checks
that occur when fusermount is trying to start up a filesystem but is not
running as root.  We'll want that for fusermount in a moment.
