+++ util/meson.build
@@ -1,18 +1,20 @@
 fuseconf_path = join_paths(get_option('prefix'), get_option('sysconfdir'), 'fuse.conf')
 
-executable('fusermount3', ['fusermount.c', '../lib/mount_util.c', '../lib/mount_fsmount.c', '../lib/util.c'],
+executable('fusermount3', ['fusermount.c', '../lib/mount_util.c', '../lib/mount_fsmount.c',
+           '../lib/util.c', 'fuser_conf.c'],
            include_directories: include_dirs,
            install: true,
            install_dir: get_option('bindir'),
            c_args: '-DFUSE_CONF="@0@"'.format(fuseconf_path))
 
 if private_cfg.get('HAVE_SERVICEMOUNT', false)
-  executable('fuservicemount3', ['mount_service.c', 'fuservicemount.c', '../lib/mount_util.c'],
+  executable('fuservicemount3', ['mount_service.c', 'fuservicemount.c', '../lib/mount_util.c',
+             'fuser_conf.c'],
              include_directories: include_dirs,
              link_with: [ libfuse ],
              install: true,
              install_dir: get_option('sbindir'),
-             c_args: '-DFUSE_USE_VERSION=319')
+             c_args: ['-DFUSE_USE_VERSION=319', '-DFUSE_CONF="@0@"'.format(fuseconf_path)])
 endif
 
 executable('mount.fuse3', ['mount.fuse.c'],
